refactor: move credentials fixtures from tests/conftest.py
Extract artifactory_setup, rhel9_http_image_url, rhsm_credentials_from_bitwarden,
and rhsm_created_secret into tests/fixtures/credentials/ (artifacts.py, rhsm.py).
Register new modules in pytest_plugins. Remove migrated fixtures and now-unused
imports from tests/conftest.py.
